Fix the EXIFAnalyzer spec
If EXIFR returns a nil value for GPS coordinates, Ruby will happily `to_f`
that to a 0.0f value. We need to actually check whether the value is nil
explicitly (or change our downstreams assumptions).

This also initializes the EXIFR logger to the Rails logger so we don’t
get stray STDERR output.