Explicitly mention house number for user street attribute
This is something I noticed while signing up myself:

As the sign up form already mentions, the address needs to be 'ladungsfähig', so the house number is most likely very relevant. : ) It is mentioned in the placeholder already, but the placeholder disappears once the user clicks into the form field. Adding it to the form label makes it a bit more bullet proof.